Pins TLS provider to < 4.0.0
Today a change was made to the Hashicorp TLS Terraform provider that
removes a field in our Terraform file. This commit pins us to a version
prior to that change.

We believe this to be reasonable because we're not testing the Terraform
process, we're testing the Ops Manager... so it doesn't matter what we
use to provision the infrastructure.

For reference, the error in question looks like this:

╷
│ Error: Invalid Configuration for Read-Only Attribute
│
│   with tls_cert_request.ssl_csr,
│   on self_signed_certs.tf line 2, in resource "tls_cert_request" "ssl_csr":
│    2:   key_algorithm   = "RSA"
│
│ Cannot set value for this attribute as the provider has marked it as
│ read-only. Remove the configuration line setting the value.
│
│ Refer to the provider documentation or contact the provider developers for
│ additional information about configurable and read-only attributes that are
│ supported.
╵

And the CI build in which this happened can probably be found at:
<https://om.ci.cf-app.com/teams/main/pipelines/opsmanager-2.10/jobs/azure-appliance-tests/builds/380>

Signed-off-by: Rajan Agaskar <ragaskar@vmware.com>